requirements.txt
ansible
pyaml